move user input logic code into lib to make it more testable
