+++ ports/servoshell/tests/mod.rs
@@ -2,7 +2,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
 
-use crate::app::parse_url_or_filename;
+use servoshell::parse_url_or_filename;
 use std::path::Path;
 
 #[cfg(not(target_os = "windows"))]
